+++ CHANGELOG.md
@@ -10,7 +10,8 @@ _..._
 ## 2.0.0 - ???
 
 - Update Plates formatter to use Payload settings for template
-- Allow Resolver to work with already instantiated objects
+- Update redirect responder to use Payload settings for target location
+- Allow resolver to work with already instantiated objects
 - Separate response formatting from HTTP status code with `StatusResponder`
 - Replace `AbstractFormatter` with a `FormatterInterface`
 - Remove dependency on Arbiter